# ImplementandoBaseDeDatos/C1_1/pkg/web/request/warehouse.go
package request

type Warehouse struct {
	Address            string `json:"address"`
	Telephone          string `json:"telephone"`
	WarehouseCode      string `json:"warehouse_code"`
	MinimumCapacity    int    `json:"minimum_capacity"`
	MinimumTemperature int    `json:"minimum_temperature"`
}